- OXListBox::MoveSelectUp/Down(): check for lbe == NULL. - OXColorSelect: conditional inclusion of the "Other..." button. - OXTextEntry: avoid having all the text scrolled to the left. - OXDockableFrame::SetAspect(): added forgotten void. - OTimer: time argument is now unsigned long. - Added the notion of insets. Added OInsets class, widgets now sould use   it instead of _bw.

+++ xclass/lib/libxclass/OTimer.cc
@@ -28,7 +28,7 @@
 
 int OTimer::id_count = 0;
 
-OTimer::OTimer(OComponent *c, int msecs) {
+OTimer::OTimer(OComponent *c, unsigned long msecs) {
 
   if (c != NULL) {
